♻️ PQAXKH task: split doctor workspace state checks
